requirements.txt
pandas
scikit-learn
notebook
scikit-surprise
matplotlib
seaborn